requirements.txt
asgiref==3.7.2
attrs==23.1.0
colorama==0.4.6
Django==4.2.5
django-js-asset==2.1.0
django-mptt==0.14.0
djangorestframework==3.14.0
drf-spectacular==0.26.4
inflection==0.5.1
iniconfig==2.0.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
packaging==23.1
pluggy==1.3.0
pytest==7.4.2
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
rpds-py==0.10.3
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
